+++ object.cpp
@@ -8,7 +8,7 @@ const sf::Color& object::get_color() const {
     return color_;
 }
 
-object::object(const sf::Color& color, const long double& reflectivity)
+object::object(const sf::Color& color, const double& reflectivity)
         : color_(color)
         , reflectivity_(reflectivity / 100)
 {}
@@ -17,6 +17,6 @@ object::~object() {
 
 }
 
-long double object::getReflectivity() const {
+double object::getReflectivity() const {
     return reflectivity_;
 }
