+++ ray_tracer.cpp
@@ -56,13 +56,13 @@ void ray_tracer::draw() const {
 }
 
 object* ray_tracer::tracey_(class ray ray, vector3d* const point) const {
-    long double min_distance = 1e18;
+    double min_distance = 1e18;
     object* intersection_object = nullptr;
 
     for(auto object : objects_) {
         vector3d intersection_point;
         if(object->intersect(ray, &intersection_point)){
-            long double current_distance = (intersection_point - ray.origin_).module();
+            double current_distance = (intersection_point - ray.origin_).module();
             if(current_distance < min_distance) {
                 min_distance = current_distance;
                 *point = intersection_point;
@@ -86,9 +86,9 @@ sf::Color ray_tracer::get_screen_color_(const vector3d& screen_point) const {
     return sf::Color();
 }
 
-long double
+double
 ray_tracer::get_point_intensity_(const vector3d& point, const object* point_object, const vector3d& dir) const {
-    long double intensity = 0.1;
+    double intensity = 0.1;
     vector3d normal = point_object->normal(point);
     if(scalar_product(normal, dir) < 0)
         normal = -normal;
@@ -101,8 +101,8 @@ ray_tracer::get_point_intensity_(const vector3d& point, const object* point_obje
 
 sf::Color ray_tracer::get_object_point_color_(const vector3d& base_point, const object* const base_object,
                                               const vector3d& direction) const {
-    long double base_intensity = get_point_intensity_(base_point, base_object, direction);
-    long double reflection = base_object->getReflectivity();
+    double base_intensity = get_point_intensity_(base_point, base_object, direction);
+    double reflection = base_object->getReflectivity();
     sf::Color base_color = base_object->get_color();
 
     if(reflection < 1e-9)
@@ -117,7 +117,7 @@ sf::Color ray_tracer::get_object_point_color_(const vector3d& base_point, const 
                                           sf::Color(),
                                           base_object->getReflectivity());
 
-    long double reflection_intensity = get_point_intensity_(reflection_point, reflection_object,
+    double reflection_intensity = get_point_intensity_(reflection_point, reflection_object,
                                                             base_point - reflection_point);
     sf::Color reflection_color = reflection_object->get_color();
 
@@ -137,16 +137,16 @@ object* ray_tracer::get_reflection_point_(const vector3d& base_point, const obje
 }
 
 sf::Color ray_tracer::calculate_reflection_dist_(const sf::Color& base_color, const sf::Color& reflection_color,
-                                                 long double reflectivity) const {
+                                                 double reflectivity) const {
     return add_(prod_(base_color, 1 - reflectivity), prod_(reflection_color, reflectivity));
 }
 
-sf::Color ray_tracer::prod_(const sf::Color& color, long double x) const {
+sf::Color ray_tracer::prod_(const sf::Color& color, double x) const {
     sf::Color result = color;
 
-    result.r = (sf::Uint8)std::min((int)(std::round(x * (long double)(result.r))), 255);
-    result.g = (sf::Uint8)std::min((int)(std::round(x * (long double)(result.g))), 255);
-    result.b = (sf::Uint8)std::min((int)(std::round(x * (long double)(result.b))), 255);
+    result.r = (sf::Uint8)std::min((int)(std::round(x * (double)(result.r))), 255);
+    result.g = (sf::Uint8)std::min((int)(std::round(x * (double)(result.g))), 255);
+    result.b = (sf::Uint8)std::min((int)(std::round(x * (double)(result.b))), 255);
 
     return result;
 }
